update register form
