+++ src/pages/register/register.module.css
@@ -102,14 +102,8 @@
   margin-bottom: 2rem;
 }
 
-.saveButton {
-  background-color: #e74c3c;
-  color: white;
-  border: none;
-  border-radius: 25px;
-  padding: 0.75rem 2rem;
-  margin-top: 2rem;
-  cursor: pointer;
+.buttonContainer {
+  justify-self: center;
 }
 
 .imageWrapper {
